fix(kernel,config): test uses f64::EPSILON (decay_rate is f64, E0308)

+++ crates/librefang-kernel/src/config.rs
@@ -804,7 +804,7 @@ mod tests {
                     .memory
                     .decay_rate)
                 .abs()
-                < f32::EPSILON
+                < f64::EPSILON
         );
     }
